pyproject.toml
[tool.poetry]
name = "data_reader"
version = "0.1.0"
description = ""
authors = ["Tafsir Thiam <ttafsir@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pluggy = "^0.13.1"
PyYAML = "^5.4.1"
click = "^7.1.2"
GitPython = "^3.1.14"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
